Simplify Kataja as a visualisation engine many things by moving free drawing into its own plugin.
