+++ kataja/Controller.py
@@ -74,7 +74,6 @@ class Controller:
         self.pressed = None  # set() # prepare for multitouch
         self.ui_pressed = None  # set() # different coordinates to pressed set
         self.text_editor_focus = None
-        self.free_drawing_mode = True
         self.dragged_focus = None
         self.dragged_text = None
         self.dragged_set = set()
