+++ kataja/actions/top_bar_actions.py
@@ -103,76 +103,6 @@ class ToggleRecording(KatajaAction):
         return ctrl.main.recorder and ctrl.main.recorder.recording
 
 
-class SwitchEditMode(KatajaAction):
-    k_action_uid = 'switch_edit_mode'
-    k_command = 'Switch editing mode'
-    k_shortcut = 'Ctrl+Shift+Space'
-    k_undoable = False
-    k_checkable = True
-    k_tooltip = '''<p><b>Visualisation:</b> Structures are created by algorithm (e.g. from 
-    plugin) and direct editing of structures is not allowed -- all changes are changes in 
-    visualisation of structure. Constituents and features are added through Lexicon panel or 
-    source files. 
-    </p>
-    <p><b>Free drawing:</b> Constituents and features can be added freely, but changes are not 
-    reflected back to syntactic model.</p> 
-    '''
-
-    def method(self, free_edit=None):
-        """ Switch between visualisation mode and free edit mode
-        :type free_edit: None to toggle between modes, True for free_drawing_mode,
-        False for visualization
-        :param state: triggering button or menu item state
-        :return:
-        """
-        if free_edit is None:
-            ctrl.free_drawing_mode = not ctrl.free_drawing_mode
-        else:
-            ctrl.free_drawing_mode = free_edit
-        ctrl.ui.update_edit_mode()
-        if ctrl.free_drawing_mode:
-            return 'Free drawing mode: draw as you will, but there is no access to derivation ' \
-                   'history for the structure.'
-        else:
-            return 'Derivation mode: you can edit the visualisation and browse the derivation ' \
-                   'history, but the underlying structure cannot be changed.'
-
-    def getter(self):
-        return not ctrl.free_drawing_mode
-
-
-class SwitchViewMode(KatajaAction):
-    k_action_uid = 'switch_view_mode'
-    k_command = 'Switch view mode'
-    k_tooltip = '''<p><b>Show all layers:</b> Show annotations and other information unnecessary 
-    for 
-    syntactic computation.</p> 
-
-    <p><b>Show only syntactic layers:</b> Nodes show only information that has effect
-     on syntactic computation.</p>
-    '''
-
-    k_shortcut = 'Shift+b'
-    k_checkable = True
-    k_undoable = False
-
-    def method(self, syntactic_mode=None):
-        """ Switch between showing only syntactic objects and showing richer representation
-        :param syntactic_mode: None to toggle between modes, True to hide other except syntactic
-        objects and values, False to show all items
-        syntactic only
-        :return:
-        """
-        if syntactic_mode is None:
-            syntactic_mode = not ctrl.doc_settings.get('syntactic_mode')
-        ctrl.forest.change_view_mode(syntactic_mode)
-
-    def getter(self):
-        return ctrl.document and ctrl.doc_settings.get('syntactic_mode')
-
-    def enabler(self):
-        return not ctrl.free_drawing_mode
-
 
 class SwitchSyntaxViewMode(KatajaAction):
     k_action_uid = 'switch_syntax_view_mode'
