+++ kataja/globals.py
@@ -113,8 +113,6 @@ USE_TRACES = 1
 TRACES_GROUPED_TOGETHER = 2
 
 # label text modes
-SYN_LABELS = 0
-SYN_LABELS_FOR_LEAVES = 1
 NODE_LABELS = 2
 NODE_LABELS_FOR_LEAVES = 3
 CHECKED_FEATURES = 5
