+++ kataja/ui_widgets/buttons/TopBarButtons.py
@@ -33,20 +33,6 @@ class TopBarButtons(QtWidgets.QFrame):
                                                 size=24,
                                                 action='toggle_recording').to_layout(layout)
 
-        self.edit_mode_button = TwoStateButton(ui_key='edit_mode_label',
-                                               parent=self,
-                                               text0='Free drawing',
-                                               text1='Visualisation',
-                                               action='switch_edit_mode').to_layout(layout)
-        layout.addStretch(0)
-
-        # Center side
-        self.view_mode_button = TwoStateButton(ui_key='view_mode_label', parent=self,
-                                               text0='Show all layers',
-                                               text1='Show only syntactic layer',
-                                               pixmap0=qt_prefs.eye_pixmap,
-                                               action='switch_view_mode').to_layout(layout)
-
         layout.addStretch(0)
 
         view_label = QtWidgets.QLabel("Visualisation:")
