+++ kataja/visualizations/DivideAndConquerTree.py
@@ -38,6 +38,7 @@ class DivideAndConquerTree(BalancedTree):
     """
     name = 'Balanced grid-based tree'
     banned_cn_shapes = (g.BRACKETED, g.SCOPEBOX)
+    use_rotation = True
 
     def __init__(self):
         BalancedTree.__init__(self)
@@ -77,15 +78,6 @@ class DivideAndConquerTree(BalancedTree):
             if node.isVisible() and (node.physics_x or node.physics_y):
                 return True
         return True
-
-    def reselect(self):
-        """ Rotate between drawing multidominated elements close to their various parents
-        """
-        self.set_data('rotation', self.get_data('rotation') - 1)
-
-    def prepare_draw(self):
-        new_rotation = self.forest.compute_traces_to_draw(self.get_data('rotation'))
-        self.set_data('rotation', new_rotation)
 
     def draw_tree(self, tree_top):
         """ Divide and conquer algorithm using a grid. Result is much like latex qtree.
