+++ plugins/FreeDrawing/setup.py
