+++ plugins/Monorail/SyntaxAPI.py
@@ -106,7 +106,7 @@ class SyntaxAPI(KatajaSyntaxAPI):
         if self.display_mode == CONSTITUENT_TREE:
             log.info('Changing to constituent tree')
             ctrl.doc_settings.set_node_setting('visible', True, g.FEATURE_NODE)
-            ctrl.doc_settings.set('label_text_mode', g.SYN_LABELS_FOR_LEAVES)
+            ctrl.doc_settings.set('label_text_mode', g.NODE_LABELS_FOR_LEAVES)
             ctrl.doc_settings.set('feature_positioning', g.HORIZONTAL_ROW)
             ctrl.doc_settings.set('feature_check_display', g.NO_CHECKING_EDGE)
             ctrl.doc_settings.set_for_edge_type('visible', False, g.CONSTITUENT_EDGE)
