+++ resources/version.txt
@@ -1,1 +1,1 @@
-2018-09-06 01:26:19.382879 | v. 18160 | alpha
+2018-09-13 00:23:39.965724 | v. 18194 | alpha
